Add card add/like/delete & popups

+++ src/utils/api.js
@@ -39,20 +39,20 @@ class Api {
         ).then((res) => this._getResponseData(res));
     }
 
-    likeCard(cardId) {
-        return fetch(`${this.url}cards/${cardId}/likes`, {
-                method: 'PUT',
-                headers: this._headers
-            }
-        ).then((res) => this._getResponseData(res))
-    }
-
-    dislikeCard(cardId) {
-        return fetch(`${this.url}cards/${cardId}/likes`, {
-                method: 'DELETE',
-                headers: this._headers
-            }
-        ).then((res) => this._getResponseData(res))
+    toggleCardLike(cardId, isLiked) {
+        if (!isLiked) {
+            return fetch(`${this.url}cards/${cardId}/likes`, {
+                    method: 'PUT',
+                    headers: this._headers
+                }
+            ).then((res) => this._getResponseData(res))
+        } else {
+            return fetch(`${this.url}cards/${cardId}/likes`, {
+                    method: 'DELETE',
+                    headers: this._headers
+                }
+            ).then((res) => this._getResponseData(res))
+        }
     }
 
     getUserInfo() {
